LogReg update
Fixed some minor issues, removed normalization of data that didn't seem
to help anything.  Using slightly larger initial theta values seems to
improve the convergence - at least it takes a while for the function to
go to NaN.  The optimizer seems to be driving h all the way to 0 and 1
values, which makes the cost function evaluate to NaN.  Interestingly,
since the gradient is still valid, the optimizer doesn't fail and seems
to continue improving the solution for a while.  I'm calling this good
for now.